main moved to sep file, begin unit tests
